[Graph]: use 32bits for vertex id.
It can reduce the space size of a graph by half, and improve
performance.

We also use 32-bit integer to count the duplicated edges in a graph.

vertex format v2.

+++ apps/graph-bfs/bfs_graph.cpp
@@ -167,5 +167,5 @@ int main(int argc, char *argv[])
 		print_io_thread_stat();
 	graph->cleanup();
 	printf("BFS from vertex %ld visits %d vertices. It takes %f seconds\n",
-			start_vertex, num_visited, time_diff(start, end));
+			(unsigned long) start_vertex, num_visited, time_diff(start, end));
 }
